requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.6
frozenlist==1.4.1
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-core==0.2.7
langchain-groq==0.1.5
langchain-text-splitters==0.2.1
langsmith==0.1.77
multidict==6.0.5
numpy==1.26.4
orjson==3.10.5
packaging==24.1
phonenumbers==8.13.39
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
python-telegram-bot==21.3
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
tenacity==8.4.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
